Use the location history model from KPublicTransport
Does the same thing, and has support for sorting by most recent use or
most often used as well as allows removal of individual entries (UI for
all of that yet to be added here though). And most importantly, it has
working de-duplication of entries.

Downside is that this loses existing cached locations.